Reorganize files
Renamed Presenter to MainPresenter and moved a few files to new "views"
and "presenters" packages.
