+++ GreekReference/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:supportsRtl="false"
         tools:ignore="UnusedAttribute">
         <activity
-            android:name="com.benlinskey.greekreference.MainActivity"
+            android:name=".views.MainActivity"
             android:label="@string/app_name"
             android:launchMode="singleTop" >
             <intent-filter>
@@ -39,17 +39,17 @@
         </activity>
         <activity
             android:name="com.benlinskey.greekreference.lexicon.LexiconDetailActivity"
-            android:parentActivityName="com.benlinskey.greekreference.MainActivity" >
+            android:parentActivityName=".views.MainActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value=".views.MainActivity" />
         </activity>
         <activity
             android:name="com.benlinskey.greekreference.syntax.SyntaxDetailActivity"
-            android:parentActivityName="com.benlinskey.greekreference.MainActivity" >
+            android:parentActivityName=".views.MainActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value=".views.MainActivity" />
         </activity>
         <activity
             android:name=".SettingsActivity"
